feat(apis): add class to interact with coverage apis
Changelog:
* Fix issue with headers in base_client
* Merge entities in a single file (temporary)
* Add class to interact with coverage apis
* Add Region entity
* Add test
