+++ navitia_client/client/base_client.py
@@ -8,4 +8,5 @@ class HttpBaseClient:
         self, auth_token: str, base_navitia_url: str = BASE_NAVITIA_URL
     ) -> None:
         self.base_navitia_url = base_navitia_url
-        self.session = Session().headers.update({"Authorization": auth_token})
+        self.session = Session()
+        self.session.headers.update({"Authorization": auth_token})
