--- navitia_client/entities/base.py
@@ -1,7 +1,0 @@
-from dataclasses import dataclass
-
-
-@dataclass
-class BaseEntity:
-    id: str
-    name: str
