--- navitia_client/entities/standard.py
@@ -1,7 +1,0 @@
-from dataclasses import dataclass
-
-
-@dataclass
-class Coord:
-    lon: float
-    lat: float
